The vehicle registration number is cross checked to verify insurance details and charge sheet of the vehicle using RPA system in highway tollbooth.
The proposed work automates toll gate supervision which allows the user to verify every vehicle for the insurance and the cases that could have been filed and inform it to the insurance provider or the Police, that arrives at the toll gate. The automation is done using UiPath Studio. It helps the police to locate the vehicles that are not insured and also helps in finding the lost vehicles.